Generate the data specifying a flametree

Generate the data specifying a flametree

flametree_grow(
  seed = 286,
  time = 6,
  scale = c(0.6, 0.8, 0.9),
  angle = c(-10, 10, 20),
  trees = 1,
  seg_col = NULL,
  seg_wid = NULL,
  shift_x = NULL,
  shift_y = NULL,
  n_shoot = NULL
)

Arguments

seed
Integer-valued seed for the random number generator

time
Number of generations to run the iterative process

scale
Vector of possible values for the "size rescaling" at each iteration

angle
Vector of possible angle changes (in degrees) at each iteration

trees
Number of trees to generate

seg_col
Function to control the segment colour

seg_wid
Function to control the segment width

shift_x
Function to control horizontal jitter

shift_y
Function to control vertical jitter

n_shoot
Function to control tree branching

Value

A tibble with the following columns: coord_x, coord_y, seg_deg, seg_len, seg_col, seg_wid, id_time, id_path, id_step, id_leaf, id_tree

The two "coord" columns specify the locations of a point. The "id" columns uniquely identify each point: id_time specifies the generation, id_path identifies each segment, and id_step contains the three values (0, 1 or 2) for the points that define each segment. The segments consist of two end points (0 and 2) and one "control" point (1) that is used to define a Bezier curve. Finally, id_tree is a numeric identifier indicating which tree the row belongs, and id_pathtree is a character variable that pastes the id_path and id_tree identifiers to create an identifier that is unique for each path in each tree.

The three "seg" columns provide summary information about each segment: seg_len is the length of the segment, seg_col is a value used to colour the segment, and seg_wid is a size parameter used to define the width of the segment

Examples

flametree_grow()

#> # A tibble: 381 x 12
#>    coord_x coord_y id_tree id_time id_path id_leaf id_pathtree id_step seg_deg
#>      <dbl>   <dbl>   <int>   <int>   <int> <lgl>   <chr>         <int>   <dbl>
#>  1   -1.11    0          1       1       1 FALSE   1_1               0      90
#>  2   -1.11    0.5        1       1       1 FALSE   1_1               1      90
#>  3   -1.11    1          1       1       1 FALSE   1_1               2      90
#>  4   -1.11    1          1       2       2 FALSE   1_2               0      80
#>  5   -1.11    1.3        1       2       2 FALSE   1_2               1      80
#>  6   -1.00    1.59       1       2       2 FALSE   1_2               2      80
#>  7   -1.11    1          1       2       3 FALSE   1_3               0     100
#>  8   -1.11    1.45       1       2       3 FALSE   1_3               1     100
#>  9   -1.26    1.89       1       2       3 FALSE   1_3               2     100
#> 10   -1.00    1.59       1       3       4 FALSE   1_4               0      90
#> # … with 371 more rows, and 3 more variables: seg_len <dbl>, seg_col <dbl>,
#> #   seg_wid <dbl>

flametree_grow(time = 10)

#> # A tibble: 6,141 x 12
#>    coord_x coord_y id_tree id_time id_path id_leaf id_pathtree id_step seg_deg
#>      <dbl>   <dbl>   <int>   <int>   <int> <lgl>   <chr>         <int>   <dbl>
#>  1    1.22    0          1       1       1 FALSE   1_1               0      90
#>  2    1.22    0.5        1       1       1 FALSE   1_1               1      90
#>  3    1.22    1          1       1       1 FALSE   1_1               2      90
#>  4    1.22    1          1       2       2 FALSE   1_2               0      80
#>  5    1.22    1.3        1       2       2 FALSE   1_2               1      80
#>  6    1.33    1.59       1       2       2 FALSE   1_2               2      80
#>  7    1.22    1          1       2       3 FALSE   1_3               0     100
#>  8    1.22    1.45       1       2       3 FALSE   1_3               1     100
#>  9    1.07    1.89       1       2       3 FALSE   1_3               2     100
#> 10    1.33    1.59       1       3       4 FALSE   1_4               0      90
#> # … with 6,131 more rows, and 3 more variables: seg_len <dbl>, seg_col <dbl>,
#> #   seg_wid <dbl>