Generate the data specifying a flametree

Generate the data specifying a flametree

flametree_grow(
  seed = 286,
  time = 6,
  scale = c(0.6, 0.8, 0.9),
  angle = c(-10, 10, 20),
  split = 2,
  prune = 0,
  trees = 1
)

Arguments

seed
Integer-valued seed for the random number generator

time
Number of generations to run the iterative process

scale
Vector of possible values for the "size rescaling" at each iteration

angle
Vector of possible angle changes (in degrees) at each iteration

split
Maximum number of shoots to generate from each tip at each iteration

prune
Probability with which a generated shoot is pruned

trees
Number of trees to generate

Value

A tibble with the following columns: coord_x, coord_y, seg_deg, seg_len, seg_col, seg_wid, id_time, id_path, id_step, id_leaf, id_tree

The two "coord" columns specify the locations of a point. The "id" columns uniquely identify each point: id_time specifies the generation, id_path identifies each segment, and id_step contains the three values (0, 1 or 2) for the points that define each segment. The segments consist of two end points (0 and 2) and one "control" point (1) that is used to define a Bezier curve. Finally, id_tree is a numeric identifier indicating which tree the row belongs, and id_pathtree is a character variable that pastes the id_path and id_tree identifiers to create an identifier that is unique for each path in each tree.

The three "seg" columns provide summary information about each segment: seg_len is the length of the segment, seg_col is a value used to colour the segment, and seg_wid is a size parameter used to define the width of the segment

Examples

flametree_grow()

#> # A tibble: 381 x 12
#>      coord_x coord_y seg_deg seg_len seg_col seg_wid id_time id_path id_step
#>        <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <int>   <int>   <int>
#>  1  0           0         90    1      0       0.955       1       1       0
#>  2  0           0.5       90    1      0.5     0.955       1       1       1
#>  3  0           1         90    1      1       0.955       1       1       2
#>  4  0           1         80    0.6    0       0.720       2       2       0
#>  5  1.84e-17    1.3       80    0.6    0.3     0.720       2       2       1
#>  6  1.04e- 1    1.59      80    0.6    0.594   0.720       2       2       2
#>  7  0           1        110    0.8    3       0.720       2       3       0
#>  8  2.45e-17    1.4      110    0.8    3.4     0.720       2       3       1
#>  9 -2.74e- 1    1.75     110    0.8    3.77    0.720       2       3       2
#> 10  1.04e- 1    1.59      70    0.48  -0.406   0.457       3       4       0
#> # … with 371 more rows, and 3 more variables: id_leaf <lgl>, id_tree <int>,
#> #   id_pathtree <chr>

flametree_grow(time = 10)

#> # A tibble: 6,141 x 12
#>      coord_x coord_y seg_deg seg_len seg_col seg_wid id_time id_path id_step
#>        <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <int>   <int>   <int>
#>  1  0           0         90    1      0       0.955       1       1       0
#>  2  0           0.5       90    1      0.5     0.955       1       1       1
#>  3  0           1         90    1      1       0.955       1       1       2
#>  4  0           1         80    0.6    0       0.720       2       2       0
#>  5  1.84e-17    1.3       80    0.6    0.3     0.720       2       2       1
#>  6  1.04e- 1    1.59      80    0.6    0.594   0.720       2       2       2
#>  7  0           1        110    0.8    3       0.720       2       3       0
#>  8  2.45e-17    1.4      110    0.8    3.4     0.720       2       3       1
#>  9 -2.74e- 1    1.75     110    0.8    3.77    0.720       2       3       2
#> 10  1.04e- 1    1.59      70    0.48  -0.406   0.457       3       4       0
#> # … with 6,131 more rows, and 3 more variables: id_leaf <lgl>, id_tree <int>,
#> #   id_pathtree <chr>